Fix all startup errors and get app running end-to-end
- mosquitto: replace invalid message_size_limit with max_packet_size, remove invalid keepalive_interval directive
- ota/models: fix BigInteger() → BigIntegerField()
- accounts, restaurants: create missing signals.py stubs
- requirements: add drf-yasg, django-health-check, mozilla-django-oidc, pyotp, qrcode, reportlab, sendgrid; unpin cryptography
- urls: guard drf_yasg, health_check, mozilla_django_oidc imports with try/except so app boots without optional packages
- reports/generators: fix unterminated string literal at line 647
- settings: add AUTH_USER_MODEL = 'accounts.User'
- analytics/models: rename related_name and db_table to avoid clash with reports.ComplianceReport
- accounts/models: add through_fields to User.restaurants M2M
- docker-compose-port5000: pass DB_* and TIMESCALEDB_* env vars so settings can connect
- entrypoint: switch from migrate --run-syncdb to migrate (migrations now exist)
- Initial migrations for all 12 local apps
- web/messages: escape inner double-quotes in no_results i18n string (en + es)

+++ services/core/apps/ota/models.py
@@ -153,7 +153,7 @@ class GatewayBackup(models.Model):
     # Backup info
     backup_path = models.CharField(max_length=255)
     version = models.CharField(max_length=50)  # Version backed up
-    size_bytes = models.BigInteger()
+    size_bytes = models.BigIntegerField()
 
     # Backup contents
     includes_database = models.BooleanField(default=True)
